+++ oauth-server/src/main/java/io/strimzi/kafka/oauth/server/ServerConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2019, Strimzi authors.
+ * Copyright 2017-2020, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 package io.strimzi.kafka.oauth.server;
@@ -22,6 +22,7 @@ public class ServerConfig extends Config {
     public static final String OAUTH_CHECK_AUDIENCE = "oauth.check.audience";
     public static final String OAUTH_CRYPTO_PROVIDER_BOUNCYCASTLE = "oauth.crypto.provider.bouncycastle";
     public static final String OAUTH_CRYPTO_PROVIDER_BOUNCYCASTLE_POSITION = "oauth.crypto.provider.bouncycastle.position";
+    public static final String OAUTH_CUSTOM_CLAIM_CHECK = "oauth.custom.claim.check";
     public static final String OAUTH_VALID_TOKEN_TYPE = "oauth.valid.token.type";
 
     public static final String STRIMZI_AUTHORIZER_DELEGATE_CLASS_NAME = "strimzi.authorizer.delegate.class.name";