+++ testsuite/client-secret-jwt-keycloak-authz-refresh-test/src/test/java/io/strimzi/testsuite/oauth/KeycloakClientCredentialsWithJwtValidationAuthzTest.java
@@ -516,7 +516,7 @@ public class KeycloakClientCredentialsWithJwtValidationAuthzTest {
 
         JsonNode token = result.get("access_token");
         if (token == null) {
-            throw new IllegalStateException("Invalid response from authorization server: no access_token");
+            throw new IllegalStateException("Invalid response from authorization server: no access_token - " + result);
         }
         return token.asText();
     }